fix layout of detail page

+++ src/resources/js/Pages/Event/EventDetail.jsx
@@ -15,20 +15,26 @@ const EventDetail = ({
     // for debug
     console.log(event, is_attended, participants, tags, wait_list);
 
+    const image = event.image || "";
+
     const [isAttended, SetIsAttended] = useState(is_attended);
     const [participantsState, SetParticipantsState] = useState(participants);
     const [waitListState, SetWaitListState] = useState(wait_list);
     const isFullyOccupied =
         participantsState.length >= event.participant_limit_number;
-    const [showAttendance, setShowAttendance] = useState(false);
-    const [showWaitList, setShowWaitList] = useState(false);
     const attendClickProcessing = useRef(false);
 
     const handleEditClick = () => {
         router.get(`/events/${event.event_id}/edit`);
     };
 
-    const handleAttendClick = async () => {
+    const handleAttendClick = async (e) => {
+        e.preventDefault();
+        e.stopPropagation();
+
+        if (attendClickProcessing.current) return;
+        attendClickProcessing.current = true;
+
         const response = await axios.post(
             `/events/${event.event_id}/participants`,
             { user_id: auth?.user?.user_id }
@@ -46,7 +52,13 @@ const EventDetail = ({
         attendClickProcessing.current = false;
     };
 
-    const handleNotAttendClick = async () => {
+    const handleNotAttendClick = async (e) => {
+        e.preventDefault();
+        e.stopPropagation();
+
+        if (attendClickProcessing.current) return;
+        attendClickProcessing.current = true;
+
         const response = await axios.delete(
             `/events/${event.event_id}/participants`,
             { user_id: auth?.user?.user_id }
@@ -64,46 +76,37 @@ const EventDetail = ({
         attendClickProcessing.current = false;
     };
 
-    console.log(waitListState);
-
-    const handleAttendanceClick = () => {
-        setShowAttendance(true);
-    };
-
-    const handleModalClose = () => {
-        setShowAttendance(false);
-        setShowWaitList(false);
-    };
-    const handleWaitListClick = () => {
-        setShowWaitList(true);
-    };
     return (
         <GeneralLayout auth={auth}>
-            <div className="flex flex-col justify-center items-center h-full pt-8 pb-6">
-                <div className="max-w-6xl w-full px-4 sm:px-6 lg:px-8">
+            <div className="flex flex-col justify-center items-center h-full pt-8">
+                <div className="max-w-7xl w-full px-4 sm:px-6 lg:px-8">
+                    <div className="flex justify-end pb-3">
+                        <button
+                            className="bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded  focus:outline-none focus:shadow-outline"
+                            onClick={handleEditClick}
+                        >
+                            Edit
+                        </button>
+                    </div>
                     <div className="bg-white overflow-hidden shadow sm:rounded-lg">
-                        <div className="h-screen max-h-96 sm:max-h-72 flex items-center">
-                            <div
-                                className={`flex bg-center bg-cover bg-no-repeat border object-contain rounded-md border-gray-200`}
-                                style={{
-                                    backgroundImage: `url(/storage/images/${event.image})`,
-                                    width: "100%",
-                                    height: "100%",
-                                }}
-                            ></div>
+                        <div className="relative">
+                            <img
+                                className="w-full h-96 object-cover sm:h-80"
+                                src={event.image}
+                                alt="Event Image"
+                            />
+                            <div className="absolute inset-0 bg-black opacity-25"></div>
                         </div>
-                        <div className="flex px-3 py-8 sm:px-10">
-                            <div className="flex flex-col w-full">
-                                <div className="flex flex-row items-center justify-between mb-4">
-                                    <h1 className="text-4xl leading-8 font-bold text-gray-900">
-                                        {event.name}
-                                    </h1>
-                                </div>
-                                <p className="text-md text-gray-600 mb-2">
-                                    Start: {event.start_date}
+                        <div className="flex px-6 py-8 sm:px-10">
+                            <div className="flex flex-col">
+                                <h1 className="text-4xl leading-8 font-bold text-gray-900 mb-4">
+                                    {event.name}
+                                </h1>
+                                <p className="text-sm text-gray-500 mb-2">
+                                    Start：{event.created_at}
                                 </p>
-                                <p className="text-md text-gray-600 mb-2">
-                                    End: {event.end_date}
+                                <p className="text-sm text-gray-500 mb-2">
+                                    End ： {event.updated_at}
                                 </p>
                             </div>
                         </div>
@@ -148,25 +151,13 @@ const EventDetail = ({
                                         </div>
                                     </dd>
                                 </div>
-
-                                <div className="sm:col-span-1 ">
+                                <div className="sm:col-span-1">
                                     <dt className="text-sm font-medium text-gray-500">
                                         Limit of Attendance
                                     </dt>
                                     <dd className="mt-1 text-sm text-gray-900">
                                         {participantsState.length}/
                                         {event.participant_limit_number}
-                                        {participantsState &&
-                                            participantsState.length > 0 && (
-                                                <button
-                                                    className="text-blue-500 ml-2 hover:underline"
-                                                    onClick={
-                                                        handleAttendanceClick
-                                                    }
-                                                >
-                                                    show attendance
-                                                </button>
-                                            )}
                                     </dd>
                                     {showAttendance && (
                                         <MemberPopup
@@ -242,7 +233,7 @@ const EventDetail = ({
                                         onClick={handleAttendClick}
                                         className="bg-orange-500 text-white rounded-md px-4 py-2 transition duration-300 ease-in-out hover:bg-orange-600"
                                     >
-                                        Waitlist
+                                        waitlist
                                     </button>
                                 </div>
                             )}
@@ -252,7 +243,7 @@ const EventDetail = ({
                                         onClick={handleAttendClick}
                                         className="bg-blue-500 text-white rounded-md px-4 py-2 transition duration-300 ease-in-out hover:bg-blue-600"
                                     >
-                                        Attend
+                                        attend
                                     </button>
                                 </div>
                             )}
@@ -262,7 +253,7 @@ const EventDetail = ({
                                         onClick={handleNotAttendClick}
                                         className="bg-red-500 text-white rounded-md px-4 py-2 transition duration-300 ease-in-out hover:bg-red-600"
                                     >
-                                        Cancel
+                                        not attend
                                     </button>
                                 </div>
                             )}
